pyproject.toml
[tool.poetry]
name = "jumpstreet"
version = "0.1.0-alpha.1"
description = "Replay system for pole vaulting"
authors = ["Spencer Hallyburton <spencer@shally.dev>"]
readme = "README.md"
repository = "https://github.com/percep-tech/jumpstreet"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
]
packages = [{include = "jumpstreet"}]

[tool.poetry.dependencies]
python = "~3.8"
numpy = "^1.19"
tqdm = "^4"
pyzmq = "^25.0"
opencv-python = ">=4.3, <4.7"
jsonpickle = "^3.0"
lib-avstack-core = {path="third_party/lib-avstack-core", develop=true}


[tool.poetry.group.test.dependencies]
pytest = "^7"
pytest-cov = "^4"
bandit = "^1.7"
black = "~22.12.0"
mypy = "^0"
isort = "^5.11"
flake8 = "^5.0"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q -s"
testpaths = [
    "tests",
    "integration",
]

[build-system]
requires = [
    "poetry-core",
]
build-backend = "poetry.core.masonry.api"